+++ CHANGELOG.md
@@ -4,6 +4,14 @@
     - extend CoreRepository
         * add datatable() function, producing datatable response
         * add orderBy() function
+        * fix getEntity() function
+    - add Criteria
+        * DateCriteria
+        * FindWhereCriteria
+        * FindWhereInCriteria
+        * LimitCriteria
+        * OrWhereCriteria
+        * Select2Criteria
 ## v.1.0.1
     - update documentation
     - add NumberTrait class